stop using selfsigned certificate on tlsroute tests
