+++ conformance/tests/backendtlspolicy.go
@@ -75,9 +75,12 @@ var BackendTLSPolicy = confsuite.ConformanceTest{
 
 			// For the re-encrypt case, we need to use the cert for the frontend tls listener.
 			certNN := types.NamespacedName{Name: "tls-validity-checks-certificate", Namespace: ns}
-			serverCertPem, _, err := GetTLSSecret(suite.Client, certNN)
+			serverCertPem, _, err := kubernetes.GetTLSSecret(suite.Client, certNN)
 			if err != nil {
 				t.Fatalf("unexpected error finding TLS secret: %v", err)
+			}
+			if len(serverCertPem) == 0 {
+				t.Fatal("missing required server certificate pem for the test")
 			}
 			// Verify that the request to a re-encrypted call to /backendTLS should succeed.
 			tls.MakeTLSRequestAndExpectEventuallyConsistentResponse(t, suite.RoundTripper, suite.TimeoutConfig, gwAddr, serverCertPem, nil, nil, "https-listener.org",
